Chorus Effect
Juno-style BBD chorus. This specializes in conjuring lush stereo pads, but also works great on leads. This is "preset-based", but there are some extra features under the hood. Some slewed noise has been used to add slight randomness to the delay times to try and mimic the imprecision of a real BBD chip. A small amount of cross-feedback is also introduced.